Change the target platforms to iOS 11.0 and tvOS 11.0 to match RN (#15)
React Native's podspec currently targets iOS 11.0 and up. Additionally, the current version of Xcode warns because this Boost podspec targets 8.0, which Xcode no longer supports. This commit changes the target version to 11.0, since 9 and 10 are unsupported by React Native anyway.

This commit also adds a changelog and changes the podspec version to 1.63.0-1. React Native pins its version of the boost-for-react-native dependency and the readme says this library's sole purpose is for React Native.

It is technically a breaking change to constrain the target platform and an earlier version of this PR bumped the CocoaPod version to 2.0.0, but decided not to take that path given the scope of this change and the fact that this library is not intended to be used outside of React Native.

Additionally, 1.63.0-1 precedes 1.63.0 according to semver, so any projects depending on `~> '1.63.0'` won't pick up this version. But to reiterate, this library's sole purpose is to be used by React Native, and the readme now explicitly says this library does not follow semver.